requirements.txt
yfinance
pandas
ta
crewai
langchain
langchain_google_genai
langchain_community
duckduckgo-search
pyTelegramBotAPI
python-dotenv